requirements.txt
# Regicide MARL Training - Unified Requirements
# Install all dependencies: pip install -r requirements.txt

# ============================================================================
# Core Dependencies
# ============================================================================
gymnasium>=0.28.0
numpy>=1.21.0

# ============================================================================
# MARL Training - PyTorch Backend (Default)
# ============================================================================
# PyTorch with CUDA support (default backend)
torch>=2.0.0

# ============================================================================
# MARL Training - JAX Backend (Optional)
# ============================================================================
# Uncomment if using JAX backend (set backend: "jax" in config)
# jax>=0.4.0
# jaxlib>=0.4.0
# chex>=0.1.7
# flax>=0.7.0
# optax>=0.1.7

# ============================================================================
# Logging and Visualization
# ============================================================================
tensorboard>=2.14.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0

# Optional: Weights & Biases for advanced experiment tracking
# wandb>=0.15.0

# ============================================================================
# Web Application (Optional - for demo/visualization)
# ============================================================================
# Uncomment if running the web interface
# Flask
# gunicorn
# psycopg2-binary 
# flask-cors  
# SQLAlchemy

# ============================================================================
# Development and Testing
# ============================================================================
pytest>=7.0.0
jupyter>=1.0.0

# ============================================================================
# Optional: JaxMARL Integration
# ============================================================================
# Install from source for official MARL implementations:
# pip install git+https://github.com/FLAIROx/JaxMARL.git